Ignore proxy classes
When using `proxy` in Clojure, it will create a new public var that is
not a `def` that doesn't have docs without an easy way to add it. It
would make more sense to doc the `def`/`defn` form that creates or
binds the `proxy` anyway. This commit adds a function to determine if
the var is a proxy and ignores it.